+++ packages/router/src/lib/provide-file-router.ts
@@ -3,12 +3,15 @@ import {
   EnvironmentProviders,
   makeEnvironmentProviders,
 } from '@angular/core';
-import { provideRouter, RouterFeatures } from '@angular/router';
+import { provideRouter, RouterFeatures, ROUTES, Routes } from '@angular/router';
+import { API_PREFIX } from '@analogjs/router/tokens';
 import { ɵHTTP_ROOT_INTERCEPTOR_FNS as HTTP_ROOT_INTERCEPTOR_FNS } from '@angular/common/http';
 
 import { routes } from './routes';
 import { updateMetaTagsOnRouteChange } from './meta-tags';
 import { cookieInterceptor } from './cookie-interceptor';
+
+declare const ANALOG_API_PREFIX: string;
 
 /**
  * Sets up providers for the Angular router, and registers
@@ -33,5 +36,22 @@ export function provideFileRouter(
       multi: true,
       useValue: cookieInterceptor,
     },
+    {
+      provide: API_PREFIX,
+      useFactory() {
+        return typeof ANALOG_API_PREFIX !== 'undefined'
+          ? ANALOG_API_PREFIX
+          : 'api';
+      },
+    },
   ]);
+}
+
+export function withExtraRoutes(routes: Routes): RouterFeatures[] {
+  return [
+    {
+      ɵkind: 100 as any,
+      ɵproviders: [{ provide: ROUTES, useValue: routes }],
+    },
+  ];
 }