+++ packages/vite-plugin-nitro/src/lib/vite-plugin-nitro.ts
@@ -62,20 +62,12 @@ export function nitro(options?: Options, nitroOptions?: NitroConfig): Plugin[] {
           additionalPagesDirs: options?.additionalPagesDirs,
         });
 
-        const apiMiddlewareHandler =
-          filePrefix +
-          normalizePath(
-            join(__dirname, `runtime/api-middleware${filePrefix ? '.mjs' : ''}`)
-          );
-        const ssrEntry = normalizePath(
-          filePrefix +
-            resolve(
-              workspaceRoot,
-              'dist',
-              rootDir,
-              `ssr/main.server${filePrefix ? '.js' : ''}`
-            )
+        const ssrEntryPath = resolve(
+          options?.ssrBuildDir ||
+            resolve(workspaceRoot, 'dist', rootDir, `ssr`),
+          `main.server${filePrefix ? '.js' : ''}`
         );
+        const ssrEntry = normalizePath(filePrefix + ssrEntryPath);
         const rendererEntry =
           filePrefix +
           normalizePath(
@@ -112,6 +104,9 @@ export function nitro(options?: Options, nitroOptions?: NitroConfig): Plugin[] {
           typescript: {
             generateTsConfig: false,
           },
+          runtimeConfig: {
+            apiPrefix: apiPrefix.substring(1),
+          },
           rollupConfig: {
             onwarn(warning) {
               if (
@@ -127,7 +122,7 @@ export function nitro(options?: Options, nitroOptions?: NitroConfig): Plugin[] {
             ...(useAPIMiddleware
               ? [
                   {
-                    handler: apiMiddlewareHandler,
+                    handler: '#ANALOG_API_MIDDLEWARE',
                     middleware: true,
                   },
                 ]
@@ -137,8 +132,37 @@ export function nitro(options?: Options, nitroOptions?: NitroConfig): Plugin[] {
           routeRules: useAPIMiddleware
             ? undefined
             : {
-                [`${apiPrefix}/**`]: { proxy: { to: '/**' } },
+                [`${apiPrefix}/**`]: {
+                  proxy: { to: '/**' },
+                },
               },
+          virtual: {
+            '#ANALOG_API_MIDDLEWARE': `
+        import { eventHandler, proxyRequest } from 'h3';
+        import { useRuntimeConfig } from '#imports';
+
+        export default eventHandler(async (event) => {
+          const apiPrefix = \`/\${useRuntimeConfig().apiPrefix}\`;
+
+          if (event.node.req.url?.startsWith(apiPrefix)) {
+            const reqUrl = event.node.req.url?.replace(apiPrefix, '');
+
+            if (
+              event.node.req.method === 'GET' &&
+              // in the case of XML routes, we want to proxy the request so that nitro gets the correct headers
+              // and can render the XML correctly as a static asset
+              !event.node.req.url?.endsWith('.xml')
+            ) {
+              return $fetch(reqUrl, { headers: event.node.req.headers });
+            }
+
+            return proxyRequest(event, reqUrl, {
+              // @ts-ignore
+              fetch: $fetch.native,
+            });
+          }
+        });`,
+          },
         };
 
         if (isVercelPreset(buildPreset)) {
@@ -244,19 +268,19 @@ export function nitro(options?: Options, nitroOptions?: NitroConfig): Plugin[] {
                * as it won't resolve the renderer.ts file correctly in node.
                */
               import { eventHandler } from 'h3';
-              
+
               // @ts-ignore
               import renderer from '${ssrEntry}';
               // @ts-ignore
               const template = \`${indexContents}\`;
-              
+
               export default eventHandler(async (event) => {
                 const html = await renderer(event.node.req.url, template, {
                   req: event.node.req,
                   res: event.node.res,
                 });
                 return html;
-              });                    
+              });
               `
               );
 
@@ -276,7 +300,7 @@ export function nitro(options?: Options, nitroOptions?: NitroConfig): Plugin[] {
                 ...(useAPIMiddleware
                   ? [
                       {
-                        handler: apiMiddlewareHandler,
+                        handler: '#ANALOG_API_MIDDLEWARE',
                         middleware: true,
                       },
                     ]
@@ -311,6 +335,11 @@ export function nitro(options?: Options, nitroOptions?: NitroConfig): Plugin[] {
               : (viteServer.config.server.host as string);
             process.env['ANALOG_PORT'] = `${viteServer.config.server.port}`;
           });
+
+          // handle upgrades if websockets are enabled
+          if (nitroOptions?.experimental?.websocket) {
+            viteServer.httpServer?.on('upgrade', server.upgrade);
+          }
 
           console.log(
             `\n\nThe server endpoints are accessible under the "${apiPrefix}" path.`
@@ -349,6 +378,16 @@ export function nitro(options?: Options, nitroOptions?: NitroConfig): Plugin[] {
             `\n\nThe '@analogjs/platform' server has been successfully built.`
           );
         }
+      },
+    },
+    {
+      name: '@analogjs/vite-plugin-nitro-api-prefix',
+      config() {
+        return {
+          define: {
+            ANALOG_API_PREFIX: `"${apiPrefix.substring(1)}"`,
+          },
+        };
       },
     },
   ];